키 수정, 서비스클래스 생성

+++ src/main/java/io/github/deogicorgi/spring/jpa/repository/TeamRepository.java
@@ -3,5 +3,5 @@ package io.github.deogicorgi.spring.jpa.repository;
 import io.github.deogicorgi.spring.jpa.entity.Team;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TeamRepository extends JpaRepository<Team, Integer> {
+public interface TeamRepository extends JpaRepository<Team, Long> {
 }
